+++ src/compas/datastructures/__init__.py
@@ -5,10 +5,6 @@ datastructures
 
 .. currentmodule:: compas.datastructures
 
-.. inheritance-diagram:: Network Mesh VolMesh
-    :parts: 1
-
-
 Meshes
 ======
 
@@ -22,17 +18,6 @@ Classes
     Mesh
 
 
-Base Classes
-------------
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
-    HalfEdge
-    BaseMesh
-
-
 Functions
 ---------
 
@@ -40,9 +25,6 @@ Functions
     :toctree: generated/
     :nosignatures:
 
-    mesh_bounding_box
-    mesh_bounding_box_xy
-    mesh_connected_components
     mesh_contours_numpy
     mesh_conway_dual
     mesh_conway_join
@@ -58,34 +40,17 @@ Functions
     mesh_conway_meta
     mesh_conway_bevel
     mesh_delete_duplicate_vertices
-    mesh_dual
     mesh_explode
-    mesh_face_adjacency
     mesh_flip_cycles
     mesh_geodesic_distances_numpy
-    mesh_is_connected
     mesh_isolines_numpy
-    mesh_merge_faces
     mesh_offset
     mesh_oriented_bounding_box_numpy
     mesh_oriented_bounding_box_xy_numpy
     mesh_planarize_faces
-    mesh_quads_to_triangles
-    mesh_slice_plane
-    mesh_smooth_centroid
-    mesh_smooth_area
-    mesh_subdivide
-    mesh_subdivide_tri
-    mesh_subdivide_corner
-    mesh_subdivide_quad
-    mesh_subdivide_catmullclark
-    mesh_subdivide_doosabin
     mesh_thicken
-    mesh_transform
-    mesh_transformed
     mesh_transform_numpy
     mesh_transformed_numpy
-    mesh_unify_cycles
     mesh_weld
     meshes_join
     meshes_join_and_weld
@@ -121,17 +86,6 @@ Classes
     Network
 
 
-Base Classes
-------------
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
-    Graph
-    BaseNetwork
-
-
 Functions
 ---------
 
@@ -164,17 +118,6 @@ Classes
     :nosignatures:
 
     VolMesh
-
-
-Base Classes
-------------
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
-    HalfFace
-    BaseVolMesh
 
 
 Functions