+++ src/compas/datastructures/mesh/__init__.py
@@ -34,6 +34,7 @@ if not IPY:
     from .pull_numpy import *  # noqa: F401 F403
     from .smoothing_numpy import *  # noqa: F401 F403
     from .transformations_numpy import *  # noqa: F401 F403
+    from .trimesh_samplepoints_numpy import *  # noqa: F401 F403
 
 from ._mesh import *  # noqa: F401 F403