+++ src/compas/robots/model/robot.py
@@ -588,7 +588,7 @@ class RobotModel(Base):
         frames = []
         for link in self.iter_links():
             if len(link.visual) and link.parent_joint:
-                frames.append(link.parent_joint.origin.copy())
+                frames.append(link.parent_joint.current_origin.copy())
         return frames
 
     @property
@@ -602,8 +602,7 @@ class RobotModel(Base):
         """
         axes = []
         for joint in self.iter_joints():
-            if joint.axis:
-                axes.append(joint.axis.vector)
+            axes.append(joint.current_axis.vector)
         return axes
 
     def __str__(self):
@@ -741,7 +740,7 @@ class RobotModel(Base):
         Frame(Point(0.000, 0.000, 0.000), Vector(0.362, 0.932, 0.000), Vector(-0.932, 0.362, 0.000))
         """
         transformations = self.compute_transformations(joint_state)
-        return [j.origin.transformed(transformations[j.name]) for j in self.iter_joints()]
+        return [j.current_origin.transformed(transformations[j.name]) for j in self.iter_joints()]
 
     def transformed_axes(self, joint_state):
         """Returns the transformed axes based on the joint_state.
@@ -765,7 +764,7 @@ class RobotModel(Base):
         [Vector(0.000, 0.000, 1.000), Vector(0.000, 0.000, 1.000)]
         """
         transformations = self.compute_transformations(joint_state)
-        return [j.axis.transformed(transformations[j.name]) for j in self.iter_joints() if j.axis.vector.length]
+        return [j.current_axis.transformed(transformations[j.name]) for j in self.iter_joints() if j.current_axis.vector.length]
 
     def forward_kinematics(self, joint_state, link_name=None):
         """Calculate the robot's forward kinematic.
@@ -798,7 +797,7 @@ class RobotModel(Base):
         joint = ee_link.parent_joint
         if joint:
             transformations = self.compute_transformations(joint_state)
-            return joint.origin.transformed(transformations[joint.name])
+            return joint.current_origin.transformed(transformations[joint.name])
         else:
             return Frame.worldXY()  # if we ask forward from base link