+++ CHANGELOG.md
@@ -14,6 +14,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 
+## [1.0.0] 2021-01-18
+
+### Added
+
+* Added `compas.datastructures.mesh.trimesh_samplepoints_numpy`.
+
+### Changed
+
+* Fix Rhino7 Mac installation path
+* Separate `compas.robots.Joint.origin` into the static parent-relative `origin` and the dynamic world-relative `current_origin`.
+* Separate `compas.robots.Joint.axis` into the static parent-relative `axis` and the dynamic world-relative `current_axis`.
+* Fixed support to convert back and forth between `compas.datastructures.Graph` and NetworkX `DiGraph`.
+
+### Removed
+
+
 ## [0.19.3] 2020-12-17
 
 ### Added